refactor(List Components): ListComponent
- refactor in preparation of the bookmark feature
  - extracted parsing logic from IndexScreen and HomeScreen
  - added ListComponent to handle selection by settings and query type
  - replaced getComponent structure in IndexScreen by ListComponent
  - replaced getListComponent structure in HomeScreen by ListComponent

SVA-62
