+++ src/components/index.js
@@ -16,6 +16,7 @@ export * from './ImagesCarousel';
 export * from './ImageTextList';
 export * from './Label';
 export * from './Link';
+export * from './ListComponent';
 export * from './ListSwitchItem';
 export * from './LoadingContainer';
 export * from './Logo';
